040-project-1/udacity-cd1756-Azure-Applications-project-main/requirements.txt
alembic==1.13.1
azure-common==1.1.28
pyopenssl==24.0.0
azure-storage-blob==12.19.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
Click==8.1.7
cryptography==41.0.5
Flask==3.0.2
Flask-Login==0.6.3
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
msal==1.3.0
pip==10.0.1
pycparser==2.21
pyodbc==5.1.0
python-dateutil==2.8.2
python-editor==1.0.4
requests==2.31.0
setuptools==58.1.0
six==1.16.0
SQLAlchemy==2.0.25
urllib3==2.2.0
Werkzeug==3.0.1
WTForms==3.1.2
python-dotenv
pyOpenSSL